Add number of uncles to aggregate statistics and fix statistics calculation
The StateProcessor was not calculating statistics correctly, treating
all stats as averages. I have separated their processing with a new
convention:

- Metrics starting with a name starting with 'Total' are simply
  aggregated per day

- Metrics starting with a name starting with 'Avg' are aggregated and
  then divided by the metric count for a given a day

Also fixed a bug with signed and non signed big integers within the
stats calculations. To be able to support reversing statistics the
big integer values must be signed.